add new push_dependencies script
delete push_sdl and push_sdlsound
