+++ version.json
@@ -1,20 +1,22 @@
 {
-    "version": "3.6.1.17",
-    "versionFriendly": "3.6.1",
-    "versionSp": "RC2",
-    "versionYear": "2024",
-    "versionMonth": "January",
-    "versionIsBeta": "true",
-    "appID": "6fcbc804-4887-4786-bcf6-b0786e1e983d",
-    "licenseLink": "https://opensource.org/license/artistic-2-0/",
-    
+  "version": "3.6.1.17",
+  "versionFriendly": "3.6.1",
+  "versionSp": "RC2",
+  "versionYear": "2024",
+  "versionMonth": "January",
+  "versionIsBeta": "true",
+  "appID": "6fcbc804-4887-4786-bcf6-b0786e1e983d",
+  "licenseLink": "https://opensource.org/license/artistic-2-0/",
+  
+  "dependencies": {
     "sdlsound": {
-        "revision": "c5639414c1bb24fb4eef5861c13adb42a4aab950",
-        "urlHash": "2e9ab24fd861f61349b9386fd66c9851e9f03a60"
+      "revision": "c5639414c1bb24fb4eef5861c13adb42a4aab950",
+      "urlHash": "2e9ab24fd861f61349b9386fd66c9851e9f03a60"
     },
     "sdl": {
-        "revision": "2.28.5",
-        "release_type": "release",
-        "urlHash": "9dba2b579b7bfbd24cb302e4f4b27563cc8f1070"
+      "revision": "2.28.5",
+      "release_type": "release",
+      "urlHash": "9dba2b579b7bfbd24cb302e4f4b27563cc8f1070"
     }
+  }
 }
